+++ compiler/rustc_parse/src/parser/mod.rs
@@ -20,7 +20,6 @@ use std::{fmt, mem, slice};
 
 use attr_wrapper::{AttrWrapper, UsePreAttrPos};
 pub use diagnostics::AttemptLocalParseRecovery;
-pub(crate) use expr::ForbiddenLetReason;
 // Public to use it for custom `if` expressions in rustfmt forks like https://github.com/tucant/rustfmt
 pub use expr::LetChainsPolicy;
 pub(crate) use item::{FnContext, FnParseMode};
